backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
ollama
PyJWT==2.8.0
email-validator==2.0.0
chromadb
bcrypt==4.1.2
python-dotenv==1.0.0
slowapi==0.1.9